Identity and Configuration now provided via two yaml files.
The image now expects the paths to two yaml files which are mounted
within the image.

1. IDENTITIES_PATH a yaml file containing all identity information
required for endpoint authentication configuration.
2. Optional CONFIG_PATH points to a yaml file that provides all configuration
information for the server, e.g. clustername, transport protocol etc.

Configuration generation reimplemented in groovy to simplify yaml
parsing and allow for more advanced templating. Furthermore, this will
enable direct use of Elytron libraries which will be required for
advanced realm maninpulation in the future.

Authentication on as default, with DIGEST supported by Hotrod and REST.

Encryption currently does not work, as there is a problem with
keystores.