Explicitly provide ubi-minimal tag
The :latest ubi-minimal tag will not work with cekit due to https://github.com/cekit/cekit/issues/637
and the underlying microdnf issue
https://bugzilla.redhat.com/show_bug.cgi?id=1769831, therefore it's
necessary to ensure that the last working tag, 8.0-213, is used.

In general it is good practice to use a specific image tag for
reproducible builds, so we should continue to explicitly provide the tag
in the future.